refactor: 코드 가독성 개선 및 데이터베이스 관련 로직을 OmokDbHelper로 이동
