+++ app/src/main/java/woowacourse/omok/MainActivity.kt
@@ -1,9 +1,6 @@
 package woowacourse.omok
 
-import android.content.ContentValues
-import android.database.Cursor
 import android.os.Bundle
-import android.util.Log
 import android.widget.ImageView
 import android.widget.TableLayout
 import android.widget.TableRow
@@ -30,29 +27,11 @@ class MainActivity : AppCompatActivity() {
     private val outputView = OutputViewAndroid()
     private val dbHelper = OmokDbHelper(this)
 
-    private fun restoreGame(boardLayout: TableLayout) {
-        val stones: List<Stone> = queryAll()
-        stones.forEach { stone ->
-            game.processTurn(stone.position, stone.color)
-            val stoneImage =
-                when (stone.color) {
-                    Color.BLACK -> R.drawable.black_stone
-                    Color.WHITE -> R.drawable.white_stone
-                }
-            boardLayout.filterImageViews().forEachIndexed { index, view ->
-                if (index % 15 == stone.position.x.value - 1 && index / 15 == stone.position.y.value - 1) {
-                    view.setImageResource(stoneImage)
-                }
-            }
-        }
-    }
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         initialize()
         val board: TableLayout = findViewById(R.id.board)
         restoreGame(board)
-
         outputView.printOmokStart(board)
         setListeners(board)
     }
@@ -65,56 +44,25 @@ class MainActivity : AppCompatActivity() {
     private fun initialize() {
         enableEdgeToEdge()
         setContentView(R.layout.activity_main)
-        ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main)) { v, insets ->
+        ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main)) { view, insets ->
             val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
-            v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
+            view.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
             insets
         }
     }
 
-    private fun insertData(
-        x: Col,
-        y: Row,
-        color: Color,
-    ) {
-        val db = dbHelper.writableDatabase
-        val values =
-            ContentValues().apply {
-                put(OmokContract.COLUMN_NAME_X, x.value)
-                put(OmokContract.COLUMN_NAME_Y, y.value)
-                put(OmokContract.COLUMN_NAME_COLOR, color.name)
-            }
-        db.insert(OmokContract.TABLE_NAME, null, values)
-        queryAll().forEach {
-            Log.i(
-                "SQL",
-                "${it.position.x.value} ${it.position.y.value} ${it.color.name}",
-            )
-        }
-        Log.i("SQL", "----------------------")
-    }
-
-    private fun queryAll(): List<Stone> {
-        val dbReader = dbHelper.readableDatabase
-        val result = mutableListOf<Stone>()
-
-        val cursor: Cursor = dbReader.rawQuery("SELECT * FROM ${OmokContract.TABLE_NAME}", null)
-
-        with(cursor) {
-            while (moveToNext()) {
-                val x: Int = getInt(getColumnIndexOrThrow(OmokContract.COLUMN_NAME_X))
-                val y: Int = getInt(getColumnIndexOrThrow(OmokContract.COLUMN_NAME_Y))
-                val color: Color =
-                    when (getString(getColumnIndexOrThrow(OmokContract.COLUMN_NAME_COLOR))) {
-                        "BLACK" -> Color.BLACK
-                        "WHITE" -> Color.WHITE
-                        else -> throw IllegalStateException()
-                    }
-                result.add(Stone(Position(Col(x), Row(y)), color))
+    private fun restoreGame(boardLayout: TableLayout) {
+        val stones: List<Stone> = dbHelper.queryAll()
+        stones.forEach { stone ->
+            game.processTurn(stone.position, stone.color)
+            boardLayout.filterImageViews().forEachIndexed { index, view ->
+                if (index % game.board.col.value == stone.position.x.value - 1 &&
+                    index / game.board.row.value == stone.position.y.value - 1
+                ) {
+                    view.setImageResource(stone.color.toImage())
+                }
             }
         }
-        cursor.close()
-        return result
     }
 
     private fun setListeners(board: TableLayout) {
@@ -129,23 +77,19 @@ class MainActivity : AppCompatActivity() {
         boardLayout: TableLayout,
     ) {
         val color: Color = game.chooseTurn()
-        val stoneImage =
-            when (color) {
-                Color.BLACK -> R.drawable.black_stone
-                Color.WHITE -> R.drawable.white_stone
-            }
-
+        val stoneImage: Int = color.toImage()
         val x = Col(index % game.board.col.value + 1)
         val y = Row(index / game.board.row.value + 1)
+
         when (val moveResult: MoveResult = game.processTurn(Position(x, y), color)) {
             is MoveResult.Success.Playing -> {
                 view.setImageResource(stoneImage)
-                insertData(x, y, color)
+                dbHelper.insertData(x, y, color)
             }
 
             is MoveResult.Success.Finished -> {
                 view.setImageResource(stoneImage)
-                insertData(x, y, color)
+                dbHelper.insertData(x, y, color)
                 outputView.printMoveResult(moveResult, this, boardLayout)
                 boardLayout.filterImageViews().forEach { it.setOnClickListener(null) }
                 dbHelper.writableDatabase.delete(OmokContract.TABLE_NAME, null, null)
@@ -158,8 +102,15 @@ class MainActivity : AppCompatActivity() {
         }
     }
 
-    private fun TableLayout.filterImageViews(): Sequence<ImageView> {
-        return children.filterIsInstance<TableRow>().flatMap { it.children }
-            .filterIsInstance<ImageView>()
-    }
+    private fun Color.toImage(): Int =
+        when (this) {
+            Color.BLACK -> R.drawable.black_stone
+            Color.WHITE -> R.drawable.white_stone
+        }
 }
+
+private fun TableLayout.filterImageViews(): Sequence<ImageView> =
+    children
+        .filterIsInstance<TableRow>()
+        .flatMap { it.children }
+        .filterIsInstance<ImageView>()
